# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Document Search / RAG API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface Document {
  id: string;
  name: string;
  status: DocumentStatus;
  chunk_count: number;
  size_bytes: number;
  file_type: string;
  created_at: string;
  /** @nullable */
  error_message?: string | null;
}

export interface DeleteResult {
  success: boolean;
  message: string;
}

export interface QueryInput {
  /** @minLength 1 */
  question: string;
  top_k?: number;
  /** Optional filter — search only these documents */
  document_ids?: string[];
}

export interface SourceChunk {
  document_id: string;
  document_name: string;
  chunk_text: string;
  score: number;
  chunk_index: number;
}

export interface QueryResult {
  answer: string;
  sources: SourceChunk[];
  /** Whether an LLM was used to generate the answer */
  has_llm: boolean;
  question: string;
  processing_time_ms: number;
}

export interface SystemStats {
  total_documents: number;
  total_chunks: number;
  ready_documents: number;
  has_llm: boolean;
  embedding_model: string;
  vector_db: string;
}