# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Intelligent Document Search / RAG API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Document management
  - name: query
    description: RAG query operations
  - name: stats
    description: System statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all uploaded documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document (PDF, DOCX, TXT) via multipart/form-data
      responses:
        "200":
          description: Document uploaded and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid file type or processing error

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a single document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Document not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document and its chunks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Document not found

  /query:
    post:
      operationId: queryDocuments
      tags: [query]
      summary: Ask a natural language question against uploaded documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryInput"
      responses:
        "200":
          description: Answer with source citations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
        "400":
          description: No documents indexed yet

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get system statistics
      responses:
        "200":
          description: System stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Document:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [processing, ready, error]
        chunk_count:
          type: integer
        size_bytes:
          type: integer
        file_type:
          type: string
        created_at:
          type: string
        error_message:
          type: ["string", "null"]
      required:
        - id
        - name
        - status
        - chunk_count
        - size_bytes
        - file_type
        - created_at

    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    QueryInput:
      type: object
      properties:
        question:
          type: string
          minLength: 1
        top_k:
          type: integer
          default: 5
        document_ids:
          type: array
          items:
            type: string
          description: Optional filter — search only these documents
      required:
        - question

    QueryResult:
      type: object
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceChunk"
        has_llm:
          type: boolean
          description: Whether an LLM was used to generate the answer
        question:
          type: string
        processing_time_ms:
          type: number
      required:
        - answer
        - sources
        - has_llm
        - question
        - processing_time_ms

    SourceChunk:
      type: object
      properties:
        document_id:
          type: string
        document_name:
          type: string
        chunk_text:
          type: string
        score:
          type: number
        chunk_index:
          type: integer
      required:
        - document_id
        - document_name
        - chunk_text
        - score
        - chunk_index

    SystemStats:
      type: object
      properties:
        total_documents:
          type: integer
        total_chunks:
          type: integer
        ready_documents:
          type: integer
        has_llm:
          type: boolean
        embedding_model:
          type: string
        vector_db:
          type: string
      required:
        - total_documents
        - total_chunks
        - ready_documents
        - has_llm
        - embedding_model
        - vector_db