# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Document Search / RAG API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from './documentStatus';

export interface Document {
  id: string;
  name: string;
  status: DocumentStatus;
  chunk_count: number;
  size_bytes: number;
  file_type: string;
  created_at: string;
  /** @nullable */
  error_message?: string | null;
}