# lib/api-zod/src/generated/types/queryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Document Search / RAG API
 * OpenAPI spec version: 0.1.0
 */

export interface QueryInput {
  /** @minLength 1 */
  question: string;
  top_k?: number;
  /** Optional filter — search only these documents */
  document_ids?: string[];
}