# lib/api-zod/src/generated/types/queryResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Document Search / RAG API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceChunk } from './sourceChunk';

export interface QueryResult {
  answer: string;
  sources: SourceChunk[];
  /** Whether an LLM was used to generate the answer */
  has_llm: boolean;
  question: string;
  processing_time_ms: number;
}