# lib/api-zod/src/generated/types/sourceChunk.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Document Search / RAG API
 * OpenAPI spec version: 0.1.0
 */

export interface SourceChunk {
  document_id: string;
  document_name: string;
  chunk_text: string;
  score: number;
  chunk_index: number;
}