# lib/api-zod/src/generated/types/systemStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Document Search / RAG API
 * OpenAPI spec version: 0.1.0
 */

export interface SystemStats {
  total_documents: number;
  total_chunks: number;
  ready_documents: number;
  has_llm: boolean;
  embedding_model: string;
  vector_db: string;
}